pyproject.toml
[tool.poetry]
name = "mqtt-playground"
version = "0.1.0"
description = "Some tests on mqtt"
authors = ["Kenji Koga <koga.k3nj1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
paho-mqtt = "^1.4"
esptool = "^2.7"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
